added basic endpoint and created sendWSRequestAndListen
